refactor(plugins): reorganize plugins package
