+++ diode_measurement/plugins/screenshot.py
@@ -3,9 +3,8 @@ import pathlib
 
 from PySide6 import QtCore, QtGui, QtWidgets
 
+from diode_measurement.core.plugin import Plugin
 from diode_measurement.utils import safe_bool
-
-from . import Plugin
 
 __all__ = ["ScreenshotPlugin"]
